testing success of command first

+++ stop_agent_containers.sh
@@ -1,4 +1,14 @@
 #!/bin/bash
 
-docker stop `docker ps --no-trunc -a -q`
-docker rm `docker ps --no-trunc -a -q`
+docker ps --no-trunc -a -q
+retval=$?
+
+# echo $retval
+
+if [ $retval -ne 0 ]; then
+	docker stop `docker ps --no-trunc -a -q`
+	docker rm `docker ps --no-trunc -a -q`
+    echo "Return code was not zero but $retval"
+fi
+
+
